pyproject.toml
[project]
name = "telefon-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adk>=0.0.5",
    "elevenlabs>=2.34.0",
    "ffmpeg>=1.4",
    "fpdf>=1.7.2",
    "google>=3.0.0",
    "google-adk>=1.24.0",
    "load-dotenv>=0.1.0",
    "praw>=7.8.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
